api/requirements.txt
python-dotenv==0.21.0
pendulum==2.1.2
redis==4.4.0
pandas==1.5.2
Flask==2.2.2
pygelf==0.4.2
